requirements.txt
asgiref==3.3.1
Django==3.1.7
django-taggit==1.3.0
Markdown==3.3.4
psycopg2==2.8.6
psycopg2-binary==2.8.6
pytz==2021.1
sqlparse==0.4.1
